refactor(ui): rename run-information-page to run-info-page
